move colors to constants
