+++ components/ClubModal/Details.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { CLUBS_GREY } from '../../colors'
+import { CLUBS_GREY } from '../../constants/colors'
 import s from 'styled-components'
 
 const Details = s.div`
