+++ components/SearchBar.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import s from 'styled-components'
 import DropdownFilter from './DropdownFilter'
-import { CLUBS_GREY } from '../colors'
+import { CLUBS_GREY } from '../constants/colors'
 
 const Wrapper = s.div`
   height: 100vh;
