+++ pages/settings.js
@@ -1,7 +1,7 @@
 import fetch from 'isomorphic-unfetch'
 import renderPage from '../renderPage.js'
 import { doApiRequest, formatResponse, API_BASE_URL, ROLE_OFFICER } from '../utils'
-import { CLUBS_GREY_LIGHT } from '../colors'
+import { CLUBS_GREY_LIGHT } from '../constants/colors'
 import { Link, Router } from '../routes'
 import React from 'react'
 import Form from '../components/Form'
